# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking packages by CPF",
        "title": "Correios API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/tracker/cpf": {
            "post": {
                "description": "Track package by CPF number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracker"
                ],
                "summary": "Track CPF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "CPF to track",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scraper.TrackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scraper.TrackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "scraper.TrackRequest": {
            "type": "object",
            "required": [
                "cpf"
            ],
            "properties": {
                "cpf": {
                    "type": "string"
                }
            }
        },
        "scraper.TrackResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/scraper.TrackingResult"
                },
                "error": {
                    "type": "string"
                },
                "scraping_method": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "scraper.TrackingEvent": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                }
            }
        },
        "scraper.TrackingResult": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/scraper.TrackingEvent"
                    }
                },
                "scraped_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}